+++ src/types/UserProjectData.ts
@@ -1,7 +1,6 @@
 export interface UserProjectData {
     org : string,
-    user: string,
     project_name : string,
     guidelines : string,
-    resources : string[],
+    resources : ProjectResource[],
 }